Abstract Navigation Menu into Desktop and Mobile components
